+++ CHANGELOG.md
@@ -12,6 +12,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 ### Security
 
+## 0.1.2 - 2021-07-20
+### Added
+- register_entity method
+- get_all_registered_service_names method
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
 ## 0.1.0 - 2021-06-10
 ### Added
 - Initialized repository